pyproject.toml
[tool.poetry]
name = "candy-shop-api"
version = "0.1.0"
description = ""
authors = ["Viktor Ashcheulov & Maksim Pautsina"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
flask-cors = "^3.0.8"
flask-marshmallow = "^0.11.0"
flask-migrate = "^2.5.3"
flask-script = "^2.0.6"
flask-sqlalchemy = "^2.4.1"
marshmallow-sqlalchemy = "^0.22.3"
sqlalchemy_utils = "^0.36.3"
python-dotenv = "^0.12.0"
gunicorn = "^20.0.4"
werkzeug = "^1.0.1"
PyJWT = "^1.7.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
